requirement.txt
streamlit==1.25.0
pillow==10.0.0